backend/food_waste_solution/requirements.txt
Django==2.2.4
requests==2.22.0
ibm_cos_sdk==2.5.2
ibm_cos_sdk_core==2.5.2
djangorestframework==3.10.2
django-cors-headers==3.0.0
